S11P12D210-94 : Feat: 수정중

+++ app/src/main/java/com/ssafy/yoganavi/data/source/lecture/LectureData.kt
@@ -1,7 +1,7 @@
 package com.ssafy.yoganavi.data.source.lecture
 
 data class LectureData(
-    val recordedId: Int,
+    val recordedId: Long,
     val recordTitle: String,
     val recordThumbnail: String,
     val likeCount: Int,
